+++ packages/auto-label/__snapshots__/auto-label.js
@@ -1,3 +1,9 @@
+exports['auto-label responds to events auto detects and labels a Spanner pull request 1'] = {
+  'labels': [
+    'api: spanner'
+  ]
+}
+
 exports['auto-label responds to events auto detects and labels a Spanner issue 1'] = {
   'labels': [
     'api: spanner'